requirements.txt
pandas==2.1.4
pydantic
loguru
sqlalchemy
psycopg2-binary
apache-airflow
pytest
